Extract the component SettingPanel
