+++ factory-ai-vision/EdgeSolution/modules/WebModule/ui_v2/src/components/SettingPanel.tsx
@@ -22,10 +22,8 @@ import {
   Checkbox,
   MessageBar,
   MessageBarType,
-  Link,
 } from '@fluentui/react';
 import { useSelector, useDispatch } from 'react-redux';
-import { useBoolean } from '@uifabric/react-hooks';
 
 import { State } from 'RootStateType';
 import {
@@ -38,12 +36,8 @@ import {
 } from '../store/setting/settingAction';
 import { WarningDialog } from './WarningDialog';
 import { dummyFunction } from '../utils/dummyFunction';
-import {
-  selectNonDemoProject,
-  createNewTrainingProject,
-  pullCVProjects,
-} from '../store/trainingProjectSlice';
-import { CreateByNameDialog } from './CreateByNameDialog';
+import { selectNonDemoProject, pullCVProjects } from '../store/trainingProjectSlice';
+import { CreateProjectDialog } from './CreateProjectDialog';
 
 type SettingPanelProps = {
   isOpen: boolean;
@@ -88,7 +82,6 @@ export const SettingPanel: React.FC<SettingPanelProps> = ({
   const [loadImgWarning, setloadImgWarning] = useState(false);
   const isCollectingData = useSelector((state: State) => state.setting.isCollectData);
   const error = useSelector((state: State) => state.setting.error);
-  const [projectDialogHidden, { setFalse: openDialg, setTrue: closeDialog }] = useBoolean(true);
 
   const dispatch = useDispatch();
 
@@ -113,10 +106,6 @@ export const SettingPanel: React.FC<SettingPanelProps> = ({
   const onLoadFullImgChange = (_, checked: boolean) => {
     if (checked) setloadImgWarning(true);
     else setLoadFullImages(checked);
-  };
-
-  const onCreateProject = async (name: string) => {
-    await dispatch(createNewTrainingProject(name));
   };
 
   const updateIsCollectData = (isCollectData, hasInit?): void => {
@@ -192,14 +181,7 @@ export const SettingPanel: React.FC<SettingPanelProps> = ({
                   selectedKey={selectedCustomvisionId}
                   calloutProps={{ calloutMaxHeight: 300 }}
                 />
-                <Link onClick={openDialg}>Create new project</Link>
-                <CreateByNameDialog
-                  hidden={projectDialogHidden}
-                  onDismiss={closeDialog}
-                  title="Create new project"
-                  subText="Create a new project will remove all the objects and images"
-                  onCreate={onCreateProject}
-                />
+                <CreateProjectDialog />
                 <Checkbox checked={loadFullImages} label="Load Full Images" onChange={onLoadFullImgChange} />
                 <WarningDialog
                   open={loadImgWarning}
